Tests/LibWeb/Text/expected/HTML/referrerPolicy-attribute.txt

link referrerPolicy initial value: ''
link referrerPolicy value after setting to "invalid": ''
link referrerPolicy value after setting to "no-referrer": 'no-referrer'
link referrerPolicy value after setting to null: ''
link referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
link referrerPolicy value after setting to "": ''
link referrerPolicy value after calling removeAttribute: ''
a referrerPolicy initial value: ''
a referrerPolicy value after setting to "invalid": ''
a referrerPolicy value after setting to "no-referrer": 'no-referrer'
a referrerPolicy value after setting to null: ''
a referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
a referrerPolicy value after setting to "": ''
a referrerPolicy value after calling removeAttribute: ''
area referrerPolicy initial value: ''
area referrerPolicy value after setting to "invalid": ''
area referrerPolicy value after setting to "no-referrer": 'no-referrer'
area referrerPolicy value after setting to null: ''
area referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
area referrerPolicy value after setting to "": ''
area referrerPolicy value after calling removeAttribute: ''
img referrerPolicy initial value: ''
img referrerPolicy value after setting to "invalid": ''
img referrerPolicy value after setting to "no-referrer": 'no-referrer'
img referrerPolicy value after setting to null: ''
img referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
img referrerPolicy value after setting to "": ''
img referrerPolicy value after calling removeAttribute: ''
iframe referrerPolicy initial value: ''
iframe referrerPolicy value after setting to "invalid": ''
iframe referrerPolicy value after setting to "no-referrer": 'no-referrer'
iframe referrerPolicy value after setting to null: ''
iframe referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
iframe referrerPolicy value after setting to "": ''
iframe referrerPolicy value after calling removeAttribute: ''
script referrerPolicy initial value: ''
script referrerPolicy value after setting to "invalid": ''
script referrerPolicy value after setting to "no-referrer": 'no-referrer'
script referrerPolicy value after setting to null: ''
script referrerPolicy value after setting to "NO-REFERRER": 'no-referrer'
script referrerPolicy value after setting to "": ''
script referrerPolicy value after calling removeAttribute: ''
